Refactored the processors from the RenderRendering pipeline into the GetRenderer pipeline. Also updated the order in which decorators are applied.
