+++ src/Extras/Jabberwocky.Extras.Polly.Sc/App_Config/Include/z.Jabberwocky/1.Jabberwocky.Extras.Polly.Sc.config
@@ -26,18 +26,12 @@
 		</events>
 
 		<pipelines>
-			<mvc.renderRendering>
-				<processor patch:before="processor[@type='Sitecore.Mvc.Pipelines.Response.RenderRendering.ExecuteRenderer, Sitecore.Mvc']"
-									 type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.RenderRendering.AddPageEditorWrapper, Jabberwocky.Extras.Polly.Sc">
-				</processor>
-				<processor patch:before="processor[@type='Sitecore.Mvc.Pipelines.Response.RenderRendering.ExecuteRenderer, Sitecore.Mvc']"
-									 type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.RenderRendering.AddPollyWrapper, Jabberwocky.Extras.Polly.Sc">
-				</processor>
-				<processor patch:before="processor[@type='Sitecore.Mvc.Pipelines.Response.RenderRendering.ExecuteRenderer, Sitecore.Mvc']"
-									 type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.RenderRendering.AddErrorHandlingWrapper, Jabberwocky.Extras.Polly.Sc">
-				</processor>
-			</mvc.renderRendering>
-		</pipelines>
+			<mvc.getRenderer>
+				<processor type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.GetRenderer.AddPollyWrapper, Jabberwocky.Extras.Polly.Sc" />
+				<processor type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.GetRenderer.AddPageEditorWrapper, Jabberwocky.Extras.Polly.Sc" />
+				<processor type="Jabberwocky.Extras.Polly.Sc.Pipelines.Mvc.GetRenderer.AddErrorHandlingWrapper, Jabberwocky.Extras.Polly.Sc" />
+			</mvc.getRenderer>
+	</pipelines>
 
 	</sitecore>
